requirements.txt
asgiref==3.8.1
colorama==0.4.6
Django==5.0.3
djangorestframework==3.16.1
django-cors-headers==4.9.0
gunicorn==21.2.0
packaging==24.0
pillow==10.3.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pytest==8.3.4
pytest-django==4.9.0
pytest-cov==6.0.0
setuptools==69.2.0
sqlparse==0.4.4
tzdata==2024.1
wheel==0.43.0
whitenoise==6.6.0
